new try

+++ hashmap.c
@@ -118,7 +118,7 @@ Pair * searchMap(HashMap * map,  char * key) {
 }
 
 Pair * firstMap(HashMap * map) {
-    for (int i = 0 ; i < map->capacity - 1 ; i++){
+    for (long i = 0 ; i < map->capacity - 1 ; i++){
         if(map -> buckets[i]!= NULL && map -> buckets[i] -> key != NULL){
             map -> current = i;
             return map -> buckets[i];
@@ -128,8 +128,8 @@ Pair * firstMap(HashMap * map) {
 }
 
 Pair * nextMap(HashMap * map) {
-    for (long i = map->current + 1 ; i < map->capacity - 1 ; i++){
-        if(map -> buckets[i]!= NULL && map -> buckets[i] -> key != NULL){
+    for (long i = map->current + 1; i < map->capacity - 1; i++){
+        if(map -> buckets[i] != NULL && map -> buckets[i] -> key != NULL){
             map -> current = i;
             return map -> buckets[i];
         }
